Updated project structure to get everything out of one file
Also added a redirect for when someone logs in with the wrong password
