+++ models.py
@@ -1,13 +1,57 @@
-from main import db
+from config import bcrypt, db
+from sqlalchemy.ext.hybrid import hybrid_property
 from datetime import datetime
 
 
 class BlogPost(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     heading = db.Column(db.String(100), unique=False, nullable=False)
+    url = db.Column(db.String(100), unique=True, nullable=False)
+    description = db.Column(db.Text, nullable=False)
     date = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
     content = db.Column(db.Text, nullable=False)
     image = db.Column(db.String(100), nullable=True)
 
     def __repr__(self):
         return '<BlogPost %r>' % self.heading
+
+
+class User(db.Model):
+    """An admin user capable of viewing reports.
+
+    :param str email: email address of user
+    :param str password: encrypted password for the user
+
+    """
+    __tablename__ = 'user'
+
+    email = db.Column(db.String, primary_key=True)
+    _password = db.Column(db.String)
+    authenticated = db.Column(db.Boolean, default=False)
+
+    @hybrid_property
+    def password(self):
+        return self._password
+
+    @password.setter
+    def password(self, plaintext):
+        self._password = bcrypt.generate_password_hash(plaintext)
+
+    def is_correct_password(self, plaintext):
+        return bcrypt.check_password_hash(self._password, plaintext)
+
+    def is_active(self):
+        """True, as all users are active."""
+        return True
+
+    def get_id(self):
+        """Return the email address to satisfy Flask-Login's requirements."""
+        return self.email
+
+    def is_authenticated(self):
+        """Return True if the user is authenticated."""
+        return self.authenticated
+
+    def is_anonymous(self):
+        """False, as anonymous users aren't supported."""
+        return False
